console: add Name function
when sending the console over i.e. (unix.Sendfd in runc), we also use name as
auxiliary data in the payload. This exposes `Name()` function. I think this
should be enough, maybe a bit leaner than exposing `File`

For windows, name is an empty string. Not sure if this is correct but i dont
know what are other options.

Signed-off-by: Daniel Dao <dqminh89@gmail.com>