Current: allow stdin to be redirected
In case we have stdin redirected, we still have a terminal which
can be obtained from stdout or stderr. This is what this commit
does.

This should allow something like this to work:

> ctr run --rm --tty docker.io/library/hello-world:latest xxx < /dev/null

NB: in case all three are redirected, but we still have a controlling
tty, we can easily get it by opening /dev/tty, but then it should be
closed, and it's not quite clear by whom and when, so this is left as
a home exersize for the reader.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>