Add support for z/OS
Add an implementation of NewPty for the zos platform. The other
functions in console_zos.go exactly match those in console_unix.go.

There is no need for ptsname/unlockpt on z/OS, so tc_zos.go looks
just like the implementation for linux/solaris except for the
omission of those functions.

Signed-off-by: Neil Johnson <najohnsn@us.ibm.com>