tc: make internal handlers take File interface
In a future patch we will split GetPty() so that users can pass a File
as the already-opened pty master, but in order for us to be able to use
the same File interface as our other external APIs we need to migrate
our internal handlers to use File rather than *os.File.

This change has no user-visible impact, as the underlying types used are
identical.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>